tested writing parquet and csv files

+++ water_column_sonar_annotation/cruise/cruise_manager.py
@@ -93,7 +93,7 @@ class CruiseManager:
             depth_min = self.get_depth(start_time, end_time)
             if np.isnan(depth_min):
                 return 0.0
-            return np.round(depth_min - bbox_max, 2).astype(np.float32)
+            return np.round(depth_min - bbox_max, 2)
         except Exception as get_altitude_exception:
             print(f"Problem getting altitude: {get_altitude_exception}")
